Merge changes. Ai progress (#57)
lots of stuff changed. added working controller to allows the controlling of when to work using the jobs set values. such as nightshift() and interval().
adding saving of values for these. add new things to IJob to support the added stuff.

* Added FORCE_STOP to EnumJobState

Used to force stop a sim working. things like missing resources for too long allows the sim to continue on with what they need to

* Started Miner Goal

* Miner Box Done

the gui is pretty much done. the basics of the tile entity is done. just need the ai to see what else we need in it

* Fixes errors on world load

* Added informative comments

* Lots of bug fixes and improvements

Sims now persist after world load.
gui now shows the proper amount of sims

* progress on ai